code cleanup + parse categories fail test

+++ src/arguments_parser.h
@@ -4,6 +4,7 @@
 #include <stdlib.h>
 #include <stdio.h>
 #include <string.h>
+#include <stdint.h>
 
 typedef struct Train {
     char* outputFile;
@@ -28,6 +29,6 @@ typedef struct Command {
     CommandData data;
 } Command;
 
-Command parse_arguments(int argc, char** argv);
+Command parse_arguments(uint8_t argc, char** argv);
 
 #endif
